requirements.txt
pandas
numpy
plotly
python-dotenv
